requirements.txt
click==8.1.3
distro==1.7.0
psutil==5.9.1
boto3==1.23.9
pycryptodomex==3.14.1
pyinstaller==5.1
requests==2.27.1
tinyaes==1.0.3
